Finish isolation of getting command ID
